SIP/no-auth.fix: More appropriate location for form API URL helper.
